Separated project into multiple files for better organization
