+++ src/utils.h
@@ -1,0 +1,14 @@
+#ifndef _UTILS_H
+#define _UTILS_H
+
+
+/* Convert a string to uppercase
+ */
+char* strtoupper(char *str);
+
+/* Trim whitespace and newlines from a string
+ */
+char* trim(char *str);
+
+
+#endif
